+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Changelog
   - `/contract/private/position-v2` Get Current Position V2 (KEYED)
   - `/contract/public/leverage-bracket` Get Current Leverage Risk Limit
   - `/contract/public/market-trade` Query the latest trade data
+  - `/account/v1/withdraw/address/list` Query Withdraw Address List (KEYED)
 - Update
   - `/contract/private/submit-order` Submit Order(SIGNED): Add new request field **stp_mode**
   - `/contract/private/order` Get Order Detail (KEYED): Add new request field **account**
@@ -21,6 +22,10 @@ Changelog
   - `/contract/private/trades` Get Order Trade (KEYED): Add new request field **account**
   - `/contract/private/transaction-history` Get Transaction History (KEYED): Add new request field **account**
   - `/contract/private/trades` Get Order Trade: Request field **symbol** changed to optional
+  - `/account/v1/currencies` Get Currencies: Add new request field **currencies**
+  - `/account/v2/deposit-withdraw/history` Get Deposit And Withdraw History (KEYED): Add new request field **startTime** and **endTime**
+  - `/spot/v2/submit_order` New Order(v2) (SIGNED): Add new request field **stpMode**
+  - `/spot/v4/batch_orders` New Batch Order(v4) (SIGNED): Add new request field **stpMode**
 
 
 ### v1.3.0 Release